Clean Code - Move RaptorSearchWindowCalculator to package 'org.opentripplanner.transit.raptor.service' from '..raptor.rangeraptor.transit'.
